requirements.txt
discord.py==1.6.0
keyboard==0.13.5